refactor: 파일 서비스 및 스키마 위치 변경
