+++ src/v1/apis/file/schemas/upload.schema.ts
